PR review - remove redundancy
